# helper_scripts/openapi_conversion/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/token/reset/own/":{"post":{"summary":"Reset Own Token","operationId":"reset_own_token_token_reset_own__post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListUserResponse"}}}}},"security":[{"OAuth2AuthorizationCodeBearer":["openid"]},{"OAuth2AuthorizationCodeBearer":[]}]}},"/token/reset/user/":{"post":{"summary":"Reset User Tokens","operationId":"reset_user_tokens_token_reset_user__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetUsersRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2AuthorizationCodeBearer":["openid"]},{"OAuth2AuthorizationCodeBearer":[]}]}},"/list_users":{"post":{"summary":"List Users","operationId":"list_users_list_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListUserQuery"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListUserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2AuthorizationCodeBearer":["openid"]},{"OAuth2AuthorizationCodeBearer":[]}]}},"/whoami":{"get":{"summary":"Whoami","operationId":"whoami_whoami_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WhoAmIResponse"}}}}},"security":[{"OAuth2AuthorizationCodeBearer":["openid"]},{"OAuth2AuthorizationCodeBearer":[]}]}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ListUserQuery":{"properties":{"query":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Query","description":"Search for users matching this query","default":""}},"type":"object","title":"ListUserQuery"},"ListUserResponse":{"properties":{"users":{"items":{"$ref":"#/components/schemas/User"},"type":"array","title":"Users"},"succes":{"type":"boolean","title":"Succes"},"detail":{"type":"string","title":"Detail"}},"type":"object","required":["users","succes","detail"],"title":"ListUserResponse"},"ResetResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"detail":{"type":"string","title":"Detail"},"resets_by_user":{"additionalProperties":{"type":"integer"},"type":"object","title":"Resets By User","description":"Map of usernames to reset counts","default":""}},"type":"object","required":["success","detail"],"title":"ResetResponse"},"ResetUsersRequest":{"properties":{"user_ids":{"items":{"type":"string"},"type":"array","title":"User Ids"}},"type":"object","required":["user_ids"],"title":"ResetUsersRequest"},"User":{"properties":{"keycloak_internal_id":{"type":"string","title":"Keycloak Internal Id"},"username":{"type":"string","title":"Username"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"firstname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Firstname"},"lastname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lastname"}},"type":"object","required":["keycloak_internal_id","username"],"title":"User"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WhoAmIResponse":{"properties":{"token":{"additionalProperties":true,"type":"object","title":"Token"},"success":{"type":"boolean","title":"Success"},"twofa_admin":{"type":"boolean","title":"Twofa Admin"}},"type":"object","required":["token","success","twofa_admin"],"title":"WhoAmIResponse"}},"securitySchemes":{"OAuth2AuthorizationCodeBearer":{"type":"oauth2","flows":{"authorizationCode":{"scopes":{},"authorizationUrl":"https://id.ys-opendesk.test.rancher.falkenstein-2.potaris.de/realms/master/login-actions/authenticate","tokenUrl":"https://id.ys-opendesk.test.rancher.falkenstein-2.potaris.de/realms/master/protocol/openid-connect/token"}}}}}}